tech-radar: improve docs
