+++ workspaces/tech-radar/plugins/tech-radar-common/src/index.ts
@@ -14,4 +14,5 @@
  * limitations under the License.
  */
 
+export * from './model';
 export * from './schema';
